Refactor and simplify the example
