+++ src/main/java/com/as12/JavaMailTemplateApplication.java
@@ -5,9 +5,7 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
 public class JavaMailTemplateApplication {
-
 	public static void main(String[] args) {
 		SpringApplication.run(JavaMailTemplateApplication.class, args);
 	}
-
 }
