+++ src/main/java/com/as12/builders/MailBuilder.java
@@ -1,88 +1,78 @@
-package com.as12.mailing;
+package com.as12.builders;
 
 import java.io.StringWriter;
 import java.util.Properties;
+
+import com.as12.models.Mail;
 import org.apache.velocity.Template;
 import org.apache.velocity.VelocityContext;
 import org.apache.velocity.app.VelocityEngine;
 
-
-public class EmailBuilder {
-
+public class MailBuilder {
 	private String subject;
 	private String mailTo;
 	private String mailFrom;
 	private String template;
-	VelocityContext velocityContext;
-	VelocityEngine velocityEngine;
+	private final VelocityContext velocityContext;
+	private final VelocityEngine velocityEngine;
 	
-	public EmailBuilder() {
+	public MailBuilder() {
 		this.mailTo = "";
 		this.mailFrom = "";
 		this.subject = "";
 		this.template = "";
 		this.velocityContext = new VelocityContext();
-		
-		// Initialize Velocity Engine
-		Properties properties = new Properties();
+		final Properties properties = new Properties();
 		properties.setProperty("input.encoding", "UTF-8");
 		properties.setProperty("output.encoding", "UTF-8");
 		properties.setProperty("resource.loader", "file, class, jar");
 		properties.setProperty("class.resource.loader.class", "org.apache.velocity.runtime.resource.loader.ClasspathResourceLoader");
 		this.velocityEngine = new VelocityEngine(properties);
-				
 	}
 	
-	public EmailBuilder Subject(String subject) {
-		
+	public MailBuilder Subject(String subject) {
 		this.subject = subject;
 		return this;
 	}
 	
-	public EmailBuilder To(String to) {
-		
+	public MailBuilder To(String to) {
 		this.mailTo = to;
 		return this;
 	}
 	
-	public EmailBuilder From(String from) {
+	public MailBuilder From(String from) {
 		this.mailFrom = from;
 		return this;
 	}
 	
-	public EmailBuilder Template(String template) {
+	public MailBuilder Template(String template) {
 		this.template = template;
 		return this;
 	}
-	public EmailBuilder AddContext(String key, String value) {
-		velocityContext.put(key, value);
-		return this;
-	}
-	public EmailBuilder AddContext(String key, Object value) {
-		velocityContext.put(key, value);
-		return this;
-	}
-	public Mail createMail() throws IllegalArgumentException { 
 
-		//Select Template
-		Template templateEngine = velocityEngine.getTemplate("templates/" + this.template);
-		
-		//Apply template
-		StringWriter stringWriter = new StringWriter();		
-		templateEngine.merge(this.velocityContext, stringWriter); 
-		
-		//Check state of the mails.
+	public MailBuilder AddContext(String key, String value) {
+		velocityContext.put(key, value);
+		return this;
+	}
+
+	public MailBuilder AddContext(String key, Object value) {
+		velocityContext.put(key, value);
+		return this;
+	}
+
+	public Mail createMail() throws IllegalArgumentException {
+		final Template templateEngine = velocityEngine.getTemplate("templates/" + this.template);
+		final StringWriter stringWriter = new StringWriter();
+		templateEngine.merge(this.velocityContext, stringWriter);
 		if(this.mailTo.isEmpty() || this.mailFrom.isEmpty()) {
 			throw new IllegalArgumentException("Missing mail headers");
 		}
-		
-		//Build mail object
-		Mail result = new Mail();
+		final Mail result = new Mail();
 		result.setMailTo(this.mailTo);
 		result.setMailFrom(this.mailFrom);
 		result.setMailContent(stringWriter.toString());
-		result.setMailSubject(this.subject); 
-		
+		result.setMailSubject(this.subject);
+
 		return result;
 	}
 }
