+++ src/main/resources/application.properties
@@ -1,15 +1,1 @@
-
-#Example configuration using Gmail SSL
-mail.protocol=smtp
-mail.host=smtp.gmail.com
-mail.port=465
-mail.smtp.socketFactory.port=465
-mail.smtp.auth=true
-mail.smtp.starttls.enable=true
-mail.smtp.debug=true
-mail.smtp.starttls.required=true
-mail.smtp.socketFactory.fallback=false
-# For real system, this should come from vault or be encrypted.
-mail.from=username@gmail.com
-mail.username=username@gmail.com
-mail.password=***********
+mail.to=recipient@gmail.com
